sockets refactor
